// frontend/src/app/swagger/model/boardReadDto.ts
/**
 * S2Retro.Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InstanceReadDto } from './instanceReadDto';


export interface BoardReadDto { 
    id?: string;
    name?: string | null;
    instances?: Array<InstanceReadDto> | null;
}